Test if connect() throw an unhandled exception
I just find an issue with connect. When called on certain server the connection failed.
It throw correctly the error but an unhandle promise is also rejected. it should not be cough by the system and should be normally trapped.
I didn't investigate much more at the moment as I spend more time to reproduce the issue and create a test case.
I will come back with more detail to understand why connecting on that server trigger the `unhandledRejection` event.